Moved caching logic into general BackendABC
